update gaussian kernel (remove constant); separate oop padding file
